(frontend) Update to 12, disable ngx-sortablejs
With the update of Angular core to version 12, this commit temporarily
suspends usage of ngx-sortablejs, of which currently no compatible
version is available. Installing it would require us to either accept
legacy peer dependencies (potentially leading to issues), using an
untrusted third-party fork, or maintaining one ourselves. As
ngx-sortablejs currently only plays a minor role in the project and no
major inconvenience is expected, we choose to simply not use it for now
and reactivate it once a compatible version becomes available.