Update all frontend deps to latest version
This commit updates all dependencies in the frontend to their latest
version, using npm-check and doing necessary migrations to prevent
breaking changes.  Note that fixing the test suite is not a priority as
of now, so the tests are likely to be (still) broken and functional
tests are done on a best-effort base.